refactor(python): split monolithic python.rs into nested submodules
Split the 5,407-line python.rs into python/ directory with surface-focused
submodules following ADR 0009:

- mod.rs: PythonParser struct, PackageParser dispatcher, re-exports
- utils.rs: shared helpers (dependency building, normalization, etc.)
- archive.rs: sdist/wheel/egg extraction, pip-origin-json parsing
- rfc822_meta.rs: PKG-INFO/METADATA RFC 822 parsing
- pyproject.rs: pyproject.toml + Poetry parsing
- setup_py.rs: AST and regex setup.py parsing
- setup_cfg.rs: setup.cfg INI parsing
- pypi_json.rs: pypi.json + pip-inspect parsing

Circular dependencies resolved by relocating cross-cutting helpers:
- extract_setup_cfg_dependency_name → utils
- pip-origin-json items → archive
- detect_pkg_info_datasource_id → utils
- apply_project_url_mappings, parse_setup_cfg_keywords → utils

Test files co-located in python/ per ADR 0009:
- python_test.rs → python/test.rs
- python_scan_test.rs → python/scan_test.rs
